# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Bouquinerie","description":"Ebook management platform","license":{"name":""},"version":"0.1.0"},"paths":{"/api/authors":{"get":{"tags":["handlers::authors"],"operationId":"list_authors","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}},{"name":"page_size","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}}],"responses":{"200":{"description":"List all authors","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorListResponse"}}}}}}},"/api/books":{"get":{"tags":["handlers::books"],"operationId":"list_books","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}},{"name":"page_size","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}}],"responses":{"200":{"description":"List all books","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookListResponse"}}}}}}},"/api/books/{id}":{"get":{"tags":["handlers::books"],"operationId":"get_book","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Book with relations","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponse"}}}},"404":{"description":"Book not found"}}}},"/api/series":{"get":{"tags":["handlers::series"],"operationId":"list_series","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}},{"name":"page_size","in":"query","required":false,"schema":{"type":["integer","null"],"format":"int64","minimum":0}}],"responses":{"200":{"description":"List all series","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeriesListResponse"}}}}}}},"/health":{"get":{"tags":["health"],"operationId":"health_check","responses":{"200":{"description":"Health check passed"},"503":{"description":"Service unhealthy"}}}}},"components":{"schemas":{"Author":{"type":"object","required":["id","name","sort_name"],"properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"sort_name":{"type":"string"}}},"AuthorListResponse":{"type":"object","required":["data","total","page","pages","page_size"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Author"}},"page":{"type":"integer","format":"int64","minimum":0},"page_size":{"type":"integer","format":"int64","minimum":0},"pages":{"type":"integer","format":"int64","minimum":0},"total":{"type":"integer","format":"int64","minimum":0}}},"Book":{"type":"object","required":["id","title","subtitle","description","language","publisher","isbn","page_count","cover_path","published_date","created_at","updated_at","sort_title"],"properties":{"cover_path":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"string","format":"uuid"},"isbn":{"type":"string"},"language":{"type":"string"},"page_count":{"type":"integer","format":"int32"},"published_date":{"type":"string","format":"date"},"publisher":{"type":"string"},"sort_title":{"type":"string"},"subtitle":{"type":"string"},"title":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"BookListResponse":{"type":"object","required":["data","total","page","pages","page_size"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Book"}},"page":{"type":"integer","format":"int64","minimum":0},"page_size":{"type":"integer","format":"int64","minimum":0},"pages":{"type":"integer","format":"int64","minimum":0},"total":{"type":"integer","format":"int64","minimum":0}}},"BookResponse":{"type":"object","required":["book","authors","series","identifiers"],"properties":{"authors":{"type":"array","items":{"$ref":"#/components/schemas/Author"}},"book":{"$ref":"#/components/schemas/Book"},"identifiers":{"type":"array","items":{"$ref":"#/components/schemas/Identifier"}},"series":{"type":"array","items":{"$ref":"#/components/schemas/Series"}}}},"Identifier":{"type":"object","required":["id","book_id","source","value"],"properties":{"book_id":{"type":"string","format":"uuid"},"id":{"type":"string","format":"uuid"},"source":{"type":"string"},"value":{"type":"string"}}},"Series":{"type":"object","required":["id","name","description","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"SeriesListResponse":{"type":"object","required":["data","total","page","pages","page_size"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Series"}},"page":{"type":"integer","format":"int64","minimum":0},"page_size":{"type":"integer","format":"int64","minimum":0},"pages":{"type":"integer","format":"int64","minimum":0},"total":{"type":"integer","format":"int64","minimum":0}}}}}}